refactor(session): 废弃旧会话入口并迁移验证覆盖;
